The Alchemy Of Finance by George Soros
"An extraordinary . . . inside look into the decision–making process of the most successful money manager of our time. Fantastic."
The Wall Street Journal

George Soros is unquestionably one of the most powerful and profitable investors in the world today. Dubbed by BusinessWeek as "the Man who Moves Markets," Soros made a fortune competing with the British pound and remains active today in the global financial community. Now, in this special edition of the classic investment book, The Alchemy of Finance, Soros presents a theoretical and practical account of current financial trends and a new paradigm by which to understand the financial market today. This edition′s expanded and revised Introduction details Soros′s innovative investment practices along with his views of the world and world order. He also describes a new paradigm for the "theory of reflexivity" which underlies his unique investment strategies. Filled with expert advice and valuable business lessons, The Alchemy of Finance reveals the timeless principles of an investing legend.